+++ libsolidity/interface/StandardCompiler.cpp
@@ -381,6 +381,7 @@ Json::Value formatImmutableReferences(map<u256, pair<string, vector<size_t>>> co
 }
 
 Json::Value collectEVMObject(
+	langutil::EVMVersion _evmVersion,
 	evmasm::LinkerObject const& _object,
 	string const* _sourceMap,
 	Json::Value _generatedSources,
@@ -392,7 +393,7 @@ Json::Value collectEVMObject(
 	if (_artifactRequested("object"))
 		output["object"] = _object.toHex();
 	if (_artifactRequested("opcodes"))
-		output["opcodes"] = evmasm::disassemble(_object.bytecode);
+		output["opcodes"] = evmasm::disassemble(_object.bytecode, _evmVersion);
 	if (_artifactRequested("sourceMap"))
 		output["sourceMap"] = _sourceMap ? *_sourceMap : "";
 	if (_artifactRequested("functionDebugData"))
@@ -1328,6 +1329,7 @@ Json::Value StandardCompiler::compileSolidity(StandardCompiler::InputsAndSetting
 			wildcardMatchesExperimental
 		))
 			evmData["bytecode"] = collectEVMObject(
+				_inputsAndSettings.evmVersion,
 				compilerStack.object(contractName),
 				compilerStack.sourceMapping(contractName),
 				compilerStack.generatedSources(contractName),
@@ -1349,6 +1351,7 @@ Json::Value StandardCompiler::compileSolidity(StandardCompiler::InputsAndSetting
 			wildcardMatchesExperimental
 		))
 			evmData["deployedBytecode"] = collectEVMObject(
+				_inputsAndSettings.evmVersion,
 				compilerStack.runtimeObject(contractName),
 				compilerStack.runtimeSourceMapping(contractName),
 				compilerStack.generatedSources(contractName, true),
@@ -1491,6 +1494,7 @@ Json::Value StandardCompiler::compileYul(InputsAndSettings _inputsAndSettings)
 			if (o.bytecode)
 				output["contracts"][sourceName][contractName]["evm"][kind] =
 					collectEVMObject(
+						_inputsAndSettings.evmVersion,
 						*o.bytecode,
 						o.sourceMappings.get(),
 						Json::arrayValue,