+++ scripts/test_antlr_grammar.sh
@@ -121,6 +121,9 @@ done < <(
       grep -v -E 'revertStatement/non_called.sol' |
       # Skipping a test with "let basefee := ..."
       grep -v -E 'inlineAssembly/basefee_berlin_function.sol' |
+      # Skipping tests with "let prevrandao := ..."
+      grep -v -E 'inlineAssembly/prevrandao_allowed_function_pre_paris.sol' |
+      grep -v -E 'inlineAssembly/prevrandao_disallowed_function_post_paris.sol' |
       # Skipping license error, unrelated to the grammar
       grep -v -E 'license/license_double5.sol' |
       grep -v -E 'license/license_hidden_unicode.sol' |