+++ test/libsolidity/syntaxTests/tryCatch/structured_spuriousDragon.sol
@@ -1,0 +1,13 @@
+contract C {
+    function f() public {
+        try this.f() {
+
+        } catch Error(string memory) {
+
+        }
+    }
+}
+// ====
+// EVMVersion: =spuriousDragon
+// ----
+// TypeError 1812: (73-112): This catch clause type cannot be used on the selected EVM version (spuriousDragon). You need at least a Byzantium-compatible EVM or use `catch { ... }`.