+++ test/libsolidity/syntaxTests/viewPureChecker/inline_assembly_instructions_allowed_london.sol
@@ -1,0 +1,11 @@
+contract C {
+    function f() public {
+        assembly {
+            pop(difficulty())
+        }
+    }
+}
+// ====
+// EVMVersion: <=london
+// ----
+// Warning 2018: (17-103): Function state mutability can be restricted to view