+++ test/libsolidity/syntaxTests/viewPureChecker/inline_assembly_instructions_disallowed_pure_paris.sol
@@ -1,0 +1,11 @@
+contract C {
+    function f() public pure {
+        assembly {
+            pop(prevrandao())
+        }
+    }
+}
+// ====
+// EVMVersion: >=paris
+// ----
+// TypeError 2527: (79-91): Function declared as pure, but this expression (potentially) reads from the environment or state and thus requires "view".